Refactoring: merging packages API into Kibana client
